refactor(dashboard): 拆出 Dashboard 头部操作栏、状态概览、最近连接与批量操作栏组件
